enosys - adding flare mainnet dex and farms
